allinone/requirements.txt
google-adk
youtube-transcript-api
python-dotenv